+++ cmd/operations/boolean.js
@@ -26,17 +26,13 @@
 // Strict mode.
 'use strict'
 
-// Standard lib.
-const path = require('path')
-
 // Local modules.
 const constants = require('../../lib/constants')
 const queue = require('../../lib/queue')
 
 // Configure.
-const options = {
+const positionals = {
   operand: {
-    coerce: path.normalize, // Positional arguments need manual normalization.
     desc: 'Path to an image file',
     normalize: true,
     type: 'string'
@@ -50,13 +46,11 @@ const options = {
 
 // Command builder.
 const builder = (yargs) => {
-  const optionNames = Object.keys(options)
   return yargs
     .strict()
     .epilog('For more information on available options, please visit https://sharp.pixelplumbing.com/api-operation#boolean')
-    .options(options)
-    .global(optionNames, false)
-    .group(optionNames, 'Command Options')
+    .positional('operand', positionals.operand)
+    .positional('operator', positionals.operator)
 }
 
 // Command handler.