+++ cmd/operations/convolve.js
@@ -26,28 +26,31 @@
 // Strict mode.
 'use strict'
 
+// Package modules.
+const pick = require('lodash.pick')
+
 // Local modules.
 const queue = require('../../lib/queue')
 
 // Configure.
 const positionals = {
-  width: {
-    desc: 'Width of the kernel in pixels',
-    type: 'number'
-  },
   height: {
     desc: 'Height of the kernel in pixels',
     type: 'number'
-  }
-}
-
-const options = {
+  },
   kernel: {
-    demandOption: true,
     desc: 'Array of length width × height containing the kernel values',
     defaultDescription: '"-1 0 1 -2 0 2 -1 0 1"',
-    type: 'string'
+    type: 'number'
   },
+  width: {
+    desc: 'Width of the kernel in pixels',
+    type: 'number'
+  }
+}
+const positionalNames = Object.keys(positionals)
+
+const options = {
   scale: {
     desc: 'The scale of the kernel in pixels',
     defaultDescription: 'sum',
@@ -59,38 +62,41 @@ const options = {
     type: 'number'
   }
 }
+const optionNames = Object.keys(options)
 
 // Command builder.
 const builder = (yargs) => {
-  const optionNames = Object.keys(options)
   return yargs
     .strict()
     .example('$0 convolve 3 3 "-1 0 1 -2 0 2 -1 0 1"', 'The output will be the convolution of the input image with the horizontal Sobel operator')
     .epilog('For more information on available options, please visit https://sharp.pixelplumbing.com/api-operation#convolve')
+    .check(argv => {
+      const length = argv.width * argv.height
+      if (!(Array.isArray(argv.kernel) && argv.kernel.length === length)) {
+        throw new Error(`Expected kernel positional to have ${length} values`)
+      }
+      return true
+    })
     .options(options)
-    .positional('width', positionals.width)
+    .positional('kernel', positionals.kernel)
     .positional('height', positionals.height)
+    .positional('width', positionals.width)
     .group(optionNames, 'Command Options')
 }
 
 // Command handler.
 const handler = (args) => {
-  const kernel = args.kernel.split(' ').map((el) => parseInt(el, 10))
-
   return queue.push(['convolve', (sharp) => {
     return sharp.convolve({
-      width: args.width,
-      height: args.height,
-      kernel,
-      scale: args.scale,
-      offset: args.offset
+      ...pick(args, positionalNames),
+      ...pick(args, optionNames)
     })
   }])
 }
 
 // Exports.
 module.exports = {
-  command: 'convolve <width> <height>',
+  command: 'convolve <width> <height> <kernel..>',
   describe: 'Convolve the image with the specified kernel',
   builder,
   handler