+++ cmd/operations/flip.js
@@ -34,6 +34,7 @@ const builder = (yargs) => {
   return yargs
     .strict()
     .epilog('For more information on available options, please visit https://sharp.pixelplumbing.com/api-operation#flip')
+    .example('$0 flip')
 }
 
 // Command handler.