+++ cmd/operations/negate.js
@@ -43,8 +43,8 @@ const builder = (yargs) => {
   return yargs
     .strict()
     .epilog('For more information on available options, please visit https://sharp.pixelplumbing.com/api-operation#negate')
+    .example('$0 negate --no-alpha')
     .options(options)
-    .global(optionNames, false)
     .group(optionNames, 'Command Options')
 }