+++ cmd/operations/recomb.js
@@ -30,29 +30,32 @@
 const queue = require('../../lib/queue')
 
 // Configure.
-const options = {
+const placeholders = {
   matrix: {
     desc: 'Recombination matrix',
-    type: 'string'
+    nargs: 3 * 3,
+    type: 'number'
   }
 }
 
 // Command builder.
 const builder = (yargs) => {
-  const optionNames = Object.keys(options)
   return yargs
     .strict()
-    .example('$0 recomb "0.3588 0.7044 0.1368 0.2990 0.5870 0.1140 0.2392 0.4696 0.0912"', 'The recomb will be applied to the output, in this case a sepia filter has been applied')
     .epilog('For more information on available options, please visit https://sharp.dimens.io/api-operation#recomb')
-    .options(options)
-    .global(optionNames, false)
-    .group(optionNames, 'Command Options')
+    .example('$0 recomb 0.3588 0.7044 0.1368 0.2990 0.5870 0.1140 0.2392 0.4696 0.0912', 'The recomb will be applied to the output, in this case a sepia filter has been applied')
+    .check(argv => {
+      if (!(Array.isArray(argv.matrix) && argv.matrix.length === 9)) {
+        throw new Error('Expected matrix positional to have 9 values')
+      }
+      return true
+    })
+    .positional('matrix', placeholders.matrix)
 }
 
 // Command handler.
 const handler = (args) => {
-  const matrix = args.matrix.split(' ').map((el) => parseFloat(el))
-
+  const { matrix } = args
   return queue.push(['recomb', (sharp) => {
     return sharp.recomb([
       matrix.slice(0, 3),
@@ -64,7 +67,7 @@ const handler = (args) => {
 
 // Exports.
 module.exports = {
-  command: 'recomb <matrix>',
+  command: 'recomb <matrix..>',
   describe: 'Recomb the image with the specified matrix',
   builder,
   handler