+++ cmd/operations/rotate.js
@@ -30,13 +30,17 @@
 const queue = require('../../lib/queue')
 
 // Configure.
-const options = {
+const positionals = {
   angle: {
-    default: 'auto',
+    defaultDescription: 'auto',
     desc: 'Angle of rotation',
     type: 'number'
-  },
+  }
+}
+
+const options = {
   background: {
+    defaultDescription: '#000000',
     desc: 'String parsed by the color module to extract values for red, green, blue and alpha',
     type: 'string'
   }
@@ -47,21 +51,16 @@ const builder = (yargs) => {
   const optionNames = Object.keys(options)
   return yargs
     .strict()
-    .example('$0 rotate', 'The output will be auto-rotated using EXIF Orientation tag')
     .epilog('For more information on available options, please visit https://sharp.pixelplumbing.com/api-operation#rotate')
+    .example('$0 rotate', 'The output will be auto-rotated using EXIF Orientation tag')
+    .positional('angle', positionals.angle)
     .options(options)
-    .global(optionNames, false)
     .group(optionNames, 'Command Options')
 }
 
 // Command handler.
 const handler = (args) => {
-  if (args.angle === 'auto') args.angle = undefined // Cast.
-  return queue.push(['rotate', (sharp) => {
-    return sharp.rotate(args.angle, {
-      background: args.background
-    })
-  }])
+  return queue.push(['rotate', (sharp) => sharp.rotate(args.angle, { background: args.background })])
 }
 
 // Exports.