+++ test/cmd/operations/convolve.js
@@ -44,22 +44,23 @@ describe('convolve', () => {
   // Default width, height, and kernel.
   const width = 3
   const height = 3
-  const kernel = '-1 0 1 -2 0 2 -1 0 1'
+  const kernel = [-1, 0, 1, -2, 0, 2, -1, 0, 1]
 
   // Reset.
   afterEach('queue', () => queue.splice(0))
   afterEach('sharp', sharp.prototype.reset)
 
-  describe('<width> <height> --kernel', () => {
+  describe('<width> <height> <kernel>', () => {
     // Run.
-    beforeEach((done) => cli.parse(['convolve', width, height, `--kernel=${kernel}`], done))
+    beforeEach((done) => cli.parse(['convolve', width, height, ...kernel], done))
 
     // Tests.
     it('must set the width, height, and kernel flags', () => {
       const args = cli.parsed.argv
       expect(args).to.have.property('width', width)
       expect(args).to.have.property('height', height)
-      expect(args).to.have.property('kernel', kernel)
+      expect(args).to.have.property('kernel')
+      expect(args.kernel).to.eql(kernel)
     })
     it('must update the pipeline', () => {
       expect(queue.pipeline).to.have.length(1)
@@ -67,23 +68,19 @@ describe('convolve', () => {
     })
     it('must execute the pipeline', () => {
       const pipeline = queue.drain(sharp())
-      sinon.assert.calledWithMatch(pipeline.convolve, {
-        width,
-        height,
-        kernel: kernel.split(' ').map((el) => parseInt(el, 10))
-      })
+      sinon.assert.calledWithMatch(pipeline.convolve, { width, height, kernel })
     })
   })
 
   describe('[options]', () => {
     describe('--offset', () => {
       // Default offset.
-      const offset = '10'
+      const offset = 10
 
-      beforeEach((done) => cli.parse(['convolve', width, height, `--kernel=${kernel}`, '--offset', offset], done))
+      beforeEach((done) => cli.parse(['convolve', width, height, ...kernel, '--offset', offset], done))
 
       it('must set the offset flag', () => {
-        expect(cli.parsed.argv).to.have.property('offset', parseInt(offset, 10))
+        expect(cli.parsed.argv).to.have.property('offset', offset)
       })
       it('must update the pipeline', () => {
         expect(queue.pipeline).to.have.length(1)
@@ -91,17 +88,17 @@ describe('convolve', () => {
       })
       it('must execute the pipeline', () => {
         const pipeline = queue.drain(sharp())
-        sinon.assert.calledWithMatch(pipeline.convolve, { offset: parseInt(offset, 10) })
+        sinon.assert.calledWithMatch(pipeline.convolve, { offset })
       })
     })
     describe('--scale', () => {
       // Default scale.
-      const scale = '10'
+      const scale = 10
 
-      beforeEach((done) => cli.parse(['convolve', width, height, `--kernel=${kernel}`, '--scale', scale], done))
+      beforeEach((done) => cli.parse(['convolve', width, height, ...kernel, '--scale', scale], done))
 
       it('must set the scale flag', () => {
-        expect(cli.parsed.argv).to.have.property('scale', parseInt(scale, 10))
+        expect(cli.parsed.argv).to.have.property('scale', scale)
       })
       it('must update the pipeline', () => {
         expect(queue.pipeline).to.have.length(1)
@@ -109,7 +106,7 @@ describe('convolve', () => {
       })
       it('must execute the pipeline', () => {
         const pipeline = queue.drain(sharp())
-        sinon.assert.calledWithMatch(pipeline.convolve, { scale: parseInt(scale, 10) })
+        sinon.assert.calledWithMatch(pipeline.convolve, { scale })
       })
     })
   })